+++ mliVec.h
@@ -6,15 +6,15 @@
 #include <stdint.h>
 
 typedef struct {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 } mliVec;
 
 mliVec mliVec_set(
-    const float x,
-    const float y,
-    const float z) {
+    const double x,
+    const double y,
+    const double z) {
     mliVec out;
     out.x = x;
     out.y = y;
@@ -48,26 +48,26 @@ mliVec mliVec_cross(
     out.z = (a.x*b.y - a.y*b.x);
     return out;}
 
-float mliVec_dot(
+double mliVec_dot(
     const mliVec a,
     const mliVec b) {
     return a.x*b.x + a.y*b.y + a.z*b.z;}
 
-mliVec mliVec_multiply(const mliVec v, const float a) {
+mliVec mliVec_multiply(const mliVec v, const double a) {
     mliVec out;
     out.x = v.x*a;
     out.y = v.y*a;
     out.z = v.z*a;
     return out;}
 
-float mliVec_norm(
+double mliVec_norm(
     const mliVec a) {
     return sqrt(mliVec_dot(a, a));}
 
 mliVec mliVec_normalized(mliVec a) {
     return mliVec_multiply(a, 1./mliVec_norm(a));}
 
-float mliVec_angle_between(const mliVec a, const mliVec b) {
+double mliVec_angle_between(const mliVec a, const mliVec b) {
     mliVec a_normalized = mliVec_multiply(a, 1./mliVec_norm(a));
     mliVec b_normalized = mliVec_multiply(b, 1./mliVec_norm(b));
     return acos(mliVec_dot(a_normalized, b_normalized));}
@@ -122,9 +122,9 @@ mliVec mliVec_mirror(
 int mliVec_equal_margin(
     const mliVec a,
     const mliVec b,
-    const float distance_margin) {
+    const double distance_margin) {
     mliVec diff;
-    float distance_squared;
+    double distance_squared;
     diff = mliVec_substract(a, b);
     distance_squared = mliVec_dot(diff, diff);
     return distance_squared <= distance_margin*distance_margin;}
